+++ modules/apps/layout/layout-theme-item-selector-web/src/main/java/com/liferay/layout/theme/item/selector/web/internal/LayoutThemeItemSelectorView.java
@@ -20,6 +20,7 @@ import com.liferay.item.selector.ItemSelectorViewDescriptorRenderer;
 import com.liferay.item.selector.criteria.UUIDItemSelectorReturnType;
 import com.liferay.layout.theme.item.selector.criterion.LayoutThemeItemSelectorCriterion;
 import com.liferay.layout.theme.item.selector.web.internal.display.context.LayoutThemeItemSelectorDisplayContext;
+import com.liferay.layout.theme.item.selector.web.internal.item.selector.LayoutThemeItemSelectorViewDescriptor;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.util.JavaConstants;
