+++ modules/apps/layout/layout-theme-item-selector-web/src/main/java/com/liferay/layout/theme/item/selector/web/internal/item/selector/LayoutThemeItemSelectorViewDescriptor.java
@@ -12,12 +12,13 @@
  * details.
  */
 
-package com.liferay.layout.theme.item.selector.web.internal;
+package com.liferay.layout.theme.item.selector.web.internal.item.selector;
 
 import com.liferay.item.selector.ItemSelectorReturnType;
 import com.liferay.item.selector.ItemSelectorViewDescriptor;
 import com.liferay.item.selector.TableItemView;
 import com.liferay.item.selector.criteria.UUIDItemSelectorReturnType;
+import com.liferay.layout.theme.item.selector.web.internal.LayoutThemeTableItemView;
 import com.liferay.layout.theme.item.selector.web.internal.display.context.LayoutThemeItemSelectorDisplayContext;
 import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.model.Theme;
